feat: allow storybook <HTML> lang and dir to be settable by pattern
The approach uses Storybook's custom head tags to inject a script into the iframe that loads the story. It interrogates the DOM for data attributes specifying the values that should be set for dir and lang attributes on the iframe's <html> element. If these aren't found then default values lang="en" dir="ltr" are set.

Because the iframe's load event is fired before the story's DOM is available, having to trigger with a timeout. An area for improvement.